requirements.txt
pandas
numpy
matplotlib
sklearn.preprocessing
mysql-connector